requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
faiss-cpu
langchain>=0.1.17
langchain-community>=0.0.24
langchain-core>=0.1.44
google-generativeai>=0.3.2
langchain-google-genai>=0.0.9
langchain-openai>=0.1.7
openai>=1.30.0
markdown-pdf
python-dotenv
pdfplumber
python-pptx
PyYAML
Pillow
pydantic>=2.6.0
tiktoken
